Podgotovka k portirovaniju na 64bit. Vykosil, gde našöl mesta, gde razmery tipov, a osobenno ukazatelej, zadavalisj cislom, a ne cerez operator sizeof(). Uvelicil razmer kuci dlä Z_Malloc do 8 mb.

+++ linuxdoom-1.10/d_net.c
@@ -89,7 +89,7 @@ doomdata_t	reboundstore;
 //
 int NetbufferSize (void)
 {
-    return (int)&(((doomdata_t *)0)->cmds[netbuffer->numtics]);
+    return (uintptr_t)&(((doomdata_t *)0)->cmds[netbuffer->numtics]);
 }
 
 //
